added session persistance by storing settings for each session id in the sqlite db
